# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Redesigned Telegram Authentication API
  version: 1.3.0
  description: |
    Authentication and user management API served by this project.
servers:
  - url: http://localhost:8081
    description: Local server
paths:
...
  /api/users:
    get:
      summary: List users
      operationId: getUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    put:
      summary: Update a user (non-password fields)
      operationId: putUsers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error
  /api/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Server error
    delete:
      summary: Delete user by ID
      operationId: deleteUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          description: Hashed password in responses; plain text only on create
        online:
          type: boolean
        files:
          type: array
          items:
            type: string
        created:
          type: integer
          format: int64
        updated:
          type: integer
          format: int64
      required: [id, name, email]
...
    UserUpdate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        online:
          type: boolean
        files:
          type: array
          items:
            type: string
      required: [id]
...